update : sojme features
