+++ indian_fertilizer/settings.py
@@ -26,6 +26,7 @@ INSTALLED_APPS = [
     'django.contrib.staticfiles',
     # Primary apps
     'shop',  
+    'authendicate',
     # Third party apps
     'jazzmin',
     'mptt',
@@ -46,7 +47,7 @@ ROOT_URLCONF = 'indian_fertilizer.urls'
 TEMPLATES = [
     {
         'BACKEND': 'django.template.backends.django.DjangoTemplates',
-        'DIRS': [ os.path.join(BASE_DIR, 'templates') ],
+        'DIRS': [ BASE_DIR / 'templates' ],
         'APP_DIRS': True,
         'OPTIONS': {
             'context_processors': [
@@ -105,18 +106,18 @@ USE_TZ = True
 # Static files (CSS, JavaScript, Images)
 STATIC_URL = 'static/'
 
-STATICFILES_DIRS = os.path.join(BASE_DIR, 'static')
+STATICFILES_DIRS = [ BASE_DIR / 'static' ]
 
 # Media Files
 MEDIA_URL = 'media/'
 
-MEDIA_ROOT = os.path.join(BASE_DIR, 'media')
+MEDIA_ROOT = [ BASE_DIR / 'media' ]
 
 # Default primary key field type
 DEFAULT_AUTO_FIELD = 'django.db.models.BigAutoField'
 
 # Set default user model
-AUTH_USER_MODEL  = 'shop.User'
+AUTH_USER_MODEL  = 'authendicate.User'
 
 # Email details
 EMAIL_BACKEND = 'django.core.mail.backends.smtp.EmailBackend'
