+++ indian_fertilizer/urls.py
@@ -5,7 +5,8 @@ from django.conf.urls.static import static
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('',include('shop.urls')),
+    path('', include('shop.urls')),
+    path('auth/', include('authendicate.urls'))
 ]
 
 if settings.DEBUG:
