+++ templates/inc/slider.html
