+++ templates/layouts/main.html
@@ -5,9 +5,12 @@
 <head>
   <meta charset="utf-8">
   <meta name="viewport" content="width=device-width, initial-scale=1">
-  <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css" rel="stylesheet"
-    integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC" crossorigin="anonymous">
+  <!-- Bootstrsap Files -->
+  <link href="{% static 'css/bootstrap.min.css' %}" rel="stylesheet">
+  <script src="{% static 'js/bootstrap.min.js' %}"></script>
+  <!-- Icon -->
   <link rel="icon" href="{% static 'images/favicon.ico.ico' %}">
+  <!-- Fontawsome CDN -->
   <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.5.1/css/all.min.css"
     integrity="sha512-DTOQO9RWCH3ppGqcWaEA1BIZOC6xxalwEsw9c2QQeAIftl+Vegovlnee1c9QX4TctnWMn13TZye+giMm8e2LwA=="
     crossorigin="anonymous" referrerpolicy="no-referrer" />
@@ -18,8 +21,7 @@
   {% endblock styles %}
 </head>
 
-<body>
-  <style>
+<style>
     *{
       padding: 0;
       margin: 0;
@@ -27,22 +29,28 @@
     body{
       background-color:rgb(207, 228, 207);
     }
+    ::-webkit-scrollbar{
+      display: none;
+    }
   </style>
-
-  {% block navbar %}
-  {% include 'shop/inc/navbar.html' %}
-  {% endblock %}
-  {% block content %}
   
-  {% endblock %}
-  {% block footer %}
-  {% include 'shop/inc/footer.html' %}
-  {% endblock %}
-  <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/js/bootstrap.bundle.min.js"
-    integrity="sha384-MrcW6ZMFYlzcLA8Nl+NtUVF0sA7MsXsP1UyJoMp4YLEuNSfAP+JcXn/tWtIaxVXM"
-    crossorigin="anonymous"></script>
 
-  <script src="{% static 'jss/index.js' %}"></script>
+<body>
+
+  <header>
+  {% include 'inc/navbar.html' %}    
+  </header>
+
+  <main>
+  {% block content %}
+  {% endblock content %}
+  </main>
+
+  <footer>
+  {% include 'inc/footer.html' %}  
+  </footer>
+  
+
   {% block scripts %}
   {% endblock scripts %}
 </body>
